Bring back code to simply "proxy" whatever is the pixel format.
From here we can re-introduce some of the thrown away code,
and simply add support for different pixel formats manually.

Should we ever find a way to convert from one input to another
properly, we can then give it another shot.
For now I don't want to spend too much time on trying to learn
the ffmpeg API and be done with a "MVP" that supports the most common
use cases.